Grammar & punctuation tweaks in README.md
I spotted (what I believe to be) some mistakes in the README file.  If any of my suggestions are unclear, please feel free to ask me to explain them.  If you generally like the suggestions but believe I've misunderstood your intent in some case, please let me know what I seem to have misunderstood and I'll be happy to revise the suggestions in my PR.  

A couple of explanations, in case they are helpful:

1. "many" versus "much" is easy to confuse if you come from a language like Spanish or Portuguese (and probably other Latin-derived languages, although I'm not certain about any others).  "Many" is for countable nouns, while "much" is for uncountable nouns.  So it is "much water" but "many glasses".
2. I tried to make the items in the list as parallel as possible, so for example I made the ending punctuation match and instead of adding "a" before a single count-noun ("a fast...template engine") I made the noun abstract and therefore uncountable ("fast...templating").